Move sample tests as part of the existing test suite
