--- .github/workflows/samples.yml
@@ -1,35 +1,0 @@
-name: Samples
-on:
-  push:
-    branches:
-      - main
-  workflow_dispatch:
-permissions:
-  contents: read
-jobs:
-  test-samples:
-    strategy:
-      matrix:
-        include:
-          - os: ubuntu-latest
-            platform: linux
-          - os: macos-latest
-            platform: darwin
-          - os: windows-latest
-            platform: win32
-    runs-on: ${{ matrix.os }}
-    name: Test Samples / ${{ matrix.platform }}
-    steps:
-      - uses: holepunchto/actions/bare-base@v1
-      - uses: actions/cache@v4
-        with:
-          path: test/samples/files
-          key: samples-${{ hashFiles('test/samples/suites/*.json') }}
-      - name: Download samples
-        run: |
-          npm i -g drives@3
-          drives mirror qm5bc1h7ooaeiiiagzaiiq7qh5ecs9ob1ufm98qr8zwsg7rauabo test/samples/files
-      - name: Verify samples
-        run: bare test/samples/scripts/verify.js
-      - name: Run sample suite
-        run: npm run test:samples
