--- samples.js
@@ -1,2 +1,0 @@
-import('./test/samples/image.js')
-import('./test/samples/video.js')
