+++ test.js
@@ -1,3 +1,8 @@
 import('./test/image')
 import('./test/video')
 import('./test/common')
+
+await import('./test/samples/scripts/download.js')
+
+import('./test/samples/image.js')
+import('./test/samples/video.js')
